+++ R/utils.R
@@ -2,7 +2,14 @@
 
 #' Order the levels of one factor by the values of another
 #'
-#' \code{order_levs()} takes two factors, the first of which has values that are
+#' @description `r lifecycle::badge("deprecated")`
+#'
+#' This function will be deprecated because it is just recapitulating
+#' functionality provided by `ggplot2::facet_grid()`. If you are using this
+#' function and don't want it it go away, please get in touch and let me know
+#' what you're using it for.
+#'
+#' `order_levs()` takes two factors, the first of which has values that are
 #' nested within the values of the second, and orders the levels of the first
 #' factor such that they are clustered within the second factor.
 #'
@@ -10,10 +17,11 @@
 #'   re-ordered.
 #'
 #' @param f1 The factor to re-order. Levels of this factor must be nested within
-#'   \code{f2}
-#' @param f2 The factor to use when re-ordering \code{f1}.
+#'   `f2`
+#' @param f2 The factor to use when re-ordering `f1`.
 #' @export
 order_levs = function(f1,...){
+    lifecycle::deprecate_warn('0.1.0', 'order_levs()')
 
     # if (is.numeric(f2)){
     #     ord = order(...)