+++ data-raw/cols_31.R
@@ -1,9 +1,0 @@
-## code to prepare `cols_31` dataset goes here
-cols_31 = rev(c("#69006f","#62e622","#cf00d0","#00a60c","#8568ff","#01e479",
-            "#ff4562","#00dda5","#bc0024","#7cffff","#630039","#daff8b",
-            "#00285d","#dbb600","#ff7fc4","#8dae00","#a7d0ff","#9e5000",
-            "#bcf2ff","#662000","#f1ffd0","#241500","#ffdfe7","#376b00",
-            "#ff8d6d","#00322c","#ff9d44","#005e75","#ffce96","#3c4400",
-            "#695000"))
-
-usethis::use_data(cols_31, overwrite = TRUE)